In this analysis, the NOAA storm database is used to examine and rank storms in terms of population health hazard and economic consequences from 1950 to 2011 in the United States.
Property and crop damage data are derived from the original data and their scaling factor.

Event types are mapped to the correct set using REGEX and the amatch function.

From the bar charts of total fatalities, injuries, and total and average property and crop damage, we draw the following conclusions:

Tornadoes cause the highest number of fatalities and injuries.
Floods are the greatest cause of property damage followed by hurricanes and tornadoes, although the most damaging on average per event are hurricanes and Storm Surge/Tide.
Drought, followed by floods and hurricanes are the cause of the greatest damage to crops, although, again, the most damaging on average per event are hurricanes.